+++ CHANGELOG.md
@@ -6,12 +6,20 @@ The format is based on `Keep a Changelog (http://keepachangelog.com/en/1.0.0/) <
 
 ## [1.3.0] - TBD
 
+### Added
+
+* Added an option 'AlwaysUseStructuredNamingConvention' to the configuration's root level. 
+  It defaults to false. 
+  If set, it activate the structured naming convention detection (and therefore the variable aggregation feature) even if it is not set in the FMU's model description.
+  
 ### Changed
 
 * If the value of an enumeration-typed variable is set, strings are now considered referring to an enumerator's name, whereas integers refer to an enumerators' value.
 * Temporarily extracted FMUs are now deleted by default
     * If there is a folder with the extracted FMU in the same directory as the FMU, this extracted content will be used and remains there after the FMU Importer stops.
     * Otherwise, the FMU is now extracted to a temporary directory. This directory and its contents are deleted after the simulation run.
+
+---
 
 ## [1.2.0] - 2024-03-28